add policy

+++ app/Http/Controllers/BranchController.php
@@ -55,6 +55,7 @@ class BranchController extends Controller
      */
     public function edit(Branch $branch)
     {
+        $this->authorize('update', $branch);
         $sections = Section::all();
         $sections = SectionResource::collection($sections)->resolve();
         $branch = BranchResource::make($branch)->resolve();
@@ -66,6 +67,7 @@ class BranchController extends Controller
      */
     public function update(UpdateRequest $request, Branch $branch)
     {
+        $this->authorize('update', $branch);
         $data = $request->validated();
         $branch->update($data);
         return redirect()->route('sections.index');
@@ -76,6 +78,7 @@ class BranchController extends Controller
      */
     public function destroy(Branch $branch)
     {
+        $this->authorize('delete', $branch);
         $branch->delete();
 
         return redirect()->route('sections.index');
